+++ src/modules/demo/DemoViewContainer.js
@@ -1,0 +1,13 @@
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import { NavigationActions } from 'react-navigation';
+import DemoView from './DemoView';
+
+export default connect(
+  () => ({}),
+  (dispatch) => {
+    return {
+      navigate: bindActionCreators(NavigationActions.navigate, dispatch),
+    };
+  },
+)(DemoView);